boundary_mask

+++ neural_lam/datastore/mllam.py
@@ -135,7 +135,9 @@ class MLLAMDatastore(BaseCartesianDatastore):
             x=slice(self._n_boundary_points, -self._n_boundary_points),
             y=slice(self._n_boundary_points, -self._n_boundary_points),
         )
-        ds_unstacked["boundary_mask"] = ds_unstacked.boundary_mask.fillna(1)
+        ds_unstacked["boundary_mask"] = ds_unstacked.boundary_mask.fillna(
+            1
+        ).astype(int)
         return self.stack_grid_coords(da_or_ds=ds_unstacked.boundary_mask)
 
     @property
